requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.7
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
pytz==2023.3.post1
selectolax==0.3.17
sniffio==1.3.0
sqlparse==0.4.4
